+++ graphics2/plotter/plot_sampler.hpp
@@ -10,6 +10,7 @@
  */
 
 #include <vector>
+#include <cmath>
 
 namespace zap { namespace graphics {
     namespace interpolators {
@@ -25,7 +26,7 @@ namespace zap { namespace graphics {
 
         template <typename T>
         T cosine(const T& A, const T& B, float parm) {
-            float val = (1.f - std::cosf(parm*float(zap::maths::PI)))*.5f;
+            float val = (1.f - std::cos(parm*float(zap::maths::PI)))*.5f;
             return (1.f - val) * A + val * B;
         }
